refactored news and persian BodySec files
